fix adding gpu environment
